remove old implementation of useList721ShopCardData, keep the new one. Remove useListShopCardData that returns relevant card data for 1155 and 721.
