+++ sdk/src/react/hooks/data/primary-sales/index.ts
@@ -5,6 +5,5 @@ export * from './useList721ShopCardData';
 export * from './useList1155SaleSupplies';
 export * from './useList1155ShopCardData';
 export * from './useListPrimarySaleItems';
-export * from './useListShopCardData';
 export * from './useShopCollectibleSaleData';
 export * from './useTokenSaleDetailsBatch';
